+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/AuthRoute.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.auth
+package com.teamoffroad.feature.auth.presentation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
