+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/navigation/AuthNavigation.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.auth.navigation
+package com.teamoffroad.feature.auth.presentation.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.teamoffroad.core.navigation.Route
-import com.teamoffroad.feature.auth.AuthRoute
+import com.teamoffroad.feature.auth.presentation.AuthRoute
 
 fun NavController.navigateAuth(navOptions: NavOptions) {
     navigate(Route.Auth, navOptions)
